вынесены повторяющиеся функции в utils и utils-mock
