+++ src/mocks/data-mock.js
@@ -1,0 +1,42 @@
+const DESCRIPTIONS = [
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.',
+
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris',
+
+  'Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus',
+];
+
+const TITLES = [
+  'Lorem ipsum dolor sit amet',
+  'consectetur adipiscing',
+  'Nullam nunc',
+];
+
+const WAYPOINT_TYPES = [
+  'Taxi',
+  'Bus',
+  'Train',
+  'Ship',
+  'Drive',
+  'Flight',
+  'Check-in',
+  'Sightseeing',
+  'Restaurant',
+];
+
+const DESTINATIONS = [
+  'Amsterdam',
+  'Chamonix',
+  'Geneva',
+  'New-York',
+  'Prague',
+  'Rio',
+  'Oslo',
+  'Moscow',
+  'Tokyo',
+  'Kioto',
+  'Berlin',
+  'Honkong',
+];
+
+export { DESCRIPTIONS, TITLES, WAYPOINT_TYPES, DESTINATIONS };
